# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for IPL Backend Service with menu management",
        "title": "IPL Backend Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/payments/billing/link": {
            "post": {
                "description": "Create a Mayar payment link for multiple billing records by IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create payment link for multiple billings",
                "parameters": [
                    {
                        "description": "Billing IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePaymentLinkMultipleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link created successfully",
                        "schema": {
                            "$ref": "#/definitions/service.PaymentLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid billing IDs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Billing not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/payments/billing/{id}/link": {
            "post": {
                "description": "Create a Mayar payment link for a billing record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create payment link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Billing ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link created successfully",
                        "schema": {
                            "$ref": "#/definitions/service.PaymentLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid billing ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Billing not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreatePaymentLinkMultipleRequest": {
            "type": "object",
            "required": [
                "billing_ids"
            ],
            "properties": {
                "billing_ids": {
                    "description": "Array of billing IDs to create payment link for",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        6,
                        2
                    ]
                }
            }
        },
...
        "service.PaymentLinkResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "billing_id": {
                    "type": "integer"
                },
                "billing_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "integer"
                },
                "invoice_id": {
                    "type": "string"
                },
                "payment_url": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
...
    }
}